+++ src/plugins/contrib/wxSmithContribItems/wxSmithKWIC/wxslcddisplay.cpp
@@ -93,7 +93,7 @@ void wxsLCDDisplay::OnBuildCreatingCode()
 			if (!ss.IsEmpty()) Codef(_T("%ASetGrayColour(%s);\n"), ss.wx_str());
 			// 6 digits is the default value.
 			if (m_iNumDigits && m_iNumDigits != 6){
-				Codef(_T("%ASetNumberDigits(%ld);\n"), m_iNumDigits);
+				Codef(_T("%ASetNumberDigits(%d);\n"), static_cast<int>(m_iNumDigits));
 			}
 			if (!m_sValue.IsEmpty()) Codef(_T("%ASetValue(wxT(\"%s\");\n"), m_sValue.wx_str());
